secure storage: change struct tee_fs_file_info member type
In struct tee_fs_file_info, the member length's origin type is size_t,
when NS user is 64 bits, and secure kernel is 32 bits, the type size_t
will have different width, which will cause xtest 20022 case fail.
This commit change the member length's type from size_t to uint64_t,
which can keep a fixed width.

Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)

+++ core/include/tee/tee_fs_key_manager.h
@@ -54,7 +54,7 @@ enum tee_fs_file_type {
 };
 
 struct tee_fs_file_info {
-	size_t length;
+	uint64_t length;
 	uint32_t backup_version_table[NUM_BLOCKS_PER_FILE / 32];
 };
